Add files via upload
**Image Captioning with Salesforce BLIP Image Captioning Large**

This project utilizes the "Salesforce/blip-image-captioning-large" model to generate textual descriptions of images. Leveraging the BERT-based transformer architecture, the model combines visual and textual data for accurate and context-rich captions. It is pre-trained on extensive datasets, making it ideal for applications like automated content generation, accessibility tools, and enhanced image search. The process involves extracting visual features with a vision encoder and generating captions with a text decoder, with continuous fine-tuning to improve performance.